Refactor file imports for consistency and update Footer component
