+++ src/App.jsx
@@ -6,11 +6,13 @@ import { Outlet } from "react-router-dom"
 export default function App() {
 
   return (
-    <>
-      <Header />
-      <Outlet />
+    <div className="min-h-screen flex flex-col">
+      <div className="flex-grow">
+        <Header />
+        <Outlet />
+      </div>
       <Footer />
-    </>
+    </div>
   )
 }
