+++ src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import Typewriter from "typewriter-effect";
-import { introdata } from "../content_option";
+import { introdata } from "../data/content_option";
 import Apenas from "../assets/Ebaa.jpeg";
 
 export default function Home() {
